feat: add more actions to entry header (#1993)
* feat: add more actions to entry header

* feat: optimize available actions filtering in MoreActions component
